update post header
The post was not rendering or in other words, it was not publishing itself. The problem was the title. The initial error-prone title was, `Building a data pipeline- uploading external data in AWS S3`. The problem was the usage of symbol **:** in the title. See this [SO post the answer by user aronisstav](https://stackoverflow.com/questions/30625044/jekyll-post-not-generated). After replacing the symbol `:` with a hyphen `-`, the post rendered itself.